+++ lib/easypost/api_key.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-# An ApiKey allows you to authenticate with the EasyPost API.
-class EasyPost::ApiKey < EasyPost::Resource
-end