+++ lib/easypost/carrier_type.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-# The CarrierType object provides an interface for determining the valid fields of a CarrierAccount.
-class EasyPost::CarrierType < EasyPost::Resource
-end