+++ lib/easypost/payload.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-# Webhook Event Payloads are triggered by changes in objects you've created via the API.
-class EasyPost::Payload < EasyPost::Resource
-end