+++ lib/easypost/postage_label.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-# PostageLabel is the object containing details about the shipping label.
-class EasyPost::PostageLabel < EasyPost::Resource
-end