+++ lib/easypost/tax_identifier.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-# TaxIdentifiers are identifying numbers or IDs that are used to charge a specific party when
-# dealing with the importing or exporting of good across international borders.
-class EasyPost::TaxIdentifier < EasyPost::Resource
-end