+++ spec/batch_spec.rb
@@ -2,14 +2,16 @@
 
 require 'spec_helper'
 
-describe EasyPost::Batch do
-  describe '.create' do
-    it 'creates a batch' do
-      batch = described_class.create(
+describe EasyPost::Services::Batch do
+  let(:client) { EasyPost::Client.new(api_key: ENV['EASYPOST_TEST_API_KEY']) }
+
+  describe 'batch service' do
+    it 'create a batch' do
+      batch = client.batch.create(
         shipments: [Fixture.one_call_buy_shipment],
       )
 
-      expect(batch).to be_an_instance_of(described_class)
+      expect(batch).to be_an_instance_of(EasyPost::Models::Batch)
       expect(batch.id).to match('batch_')
       expect(batch.shipments).not_to be_nil
     end
@@ -17,19 +19,19 @@ describe EasyPost::Batch do
 
   describe '.retrieve' do
     it 'retrieves a batch' do
-      batch = described_class.create(
+      batch = client.batch.create(
         shipments: [Fixture.one_call_buy_shipment],
       )
-      retrieved_batch = described_class.retrieve(batch.id)
+      retrieved_batch = client.batch.retrieve(batch.id)
 
-      expect(retrieved_batch).to be_an_instance_of(described_class)
+      expect(retrieved_batch).to be_an_instance_of(EasyPost::Models::Batch)
       expect(retrieved_batch.id).to eq(batch.id)
     end
   end
 
   describe '.all' do
     it 'retrieves all batches' do
-      batches = described_class.all(
+      batches = client.batch.all(
         page_size: Fixture.page_size,
       )
 
@@ -37,20 +39,20 @@ describe EasyPost::Batch do
 
       expect(batch_array.count).to be <= Fixture.page_size
       expect(batches.has_more).not_to be_nil
-      expect(batch_array).to all(be_an_instance_of(described_class))
+      expect(batch_array).to all(be_an_instance_of(EasyPost::Models::Batch))
     end
   end
 
   describe '.create_and_buy' do
     it 'creates and buys a batch in a single call' do
-      batch = described_class.create_and_buy(
+      batch = client.batch.create_and_buy(
         shipments: [
           Fixture.one_call_buy_shipment,
           Fixture.one_call_buy_shipment,
         ],
       )
 
-      expect(batch).to be_an_instance_of(described_class)
+      expect(batch).to be_an_instance_of(EasyPost::Models::Batch)
       expect(batch.id).to match('batch_')
       expect(batch.num_shipments).to eq(2)
     end
@@ -58,72 +60,72 @@ describe EasyPost::Batch do
 
   describe '.buy' do
     it 'buys a batch' do
-      batch = described_class.create(
+      batch = client.batch.create(
         shipments: [Fixture.one_call_buy_shipment],
       )
 
-      batch.buy
+      bought_batch = client.batch.buy(batch.id)
 
-      expect(batch).to be_an_instance_of(described_class)
-      expect(batch.num_shipments).to eq(1)
+      expect(bought_batch).to be_an_instance_of(EasyPost::Models::Batch)
+      expect(bought_batch.num_shipments).to eq(1)
     end
   end
 
   describe '.create_scan_form' do
     it 'creates a scanform for a batch', :vcr do
-      batch = described_class.create(
+      batch = client.batch.create(
         shipments: [Fixture.one_call_buy_shipment],
       )
 
-      batch.buy
+      bought_batch = client.batch.buy(batch.id)
 
       unless File.file?(VCR.current_cassette.file)
         sleep(5) # Wait enough time for the batch to process buying the shipment
       end
 
-      batch.create_scan_form
+      batch_with_scan_form = client.batch.create_scan_form(bought_batch.id)
 
       # We can't assert anything meaningful here because the scanform gets queued for generation and may not be immediately available
-      expect(batch).to be_an_instance_of(described_class)
+      expect(batch_with_scan_form).to be_an_instance_of(EasyPost::Models::Batch)
     end
   end
 
-  describe 'add/remove shipments' do
+  describe 'add and remove shipments' do
     it 'adds and removes shipments from a batch' do
-      shipment = EasyPost::Shipment.create(Fixture.one_call_buy_shipment)
+      shipment = client.shipment.create(Fixture.one_call_buy_shipment)
 
-      batch = described_class.create
+      batch = client.batch.create
 
-      batch.add_shipments(
-        shipments: [shipment],
+      batch_with_shipment = client.batch.add_shipments(
+        batch.id,
+        shipment: [{ id: shipment.id }],
       )
-      expect(batch.num_shipments).to eq(1)
+      expect(batch_with_shipment.num_shipments).to eq(1)
 
-      batch.remove_shipments(
-        shipments: [shipment],
+      batch_without_shipment = client.batch.remove_shipments(
+        batch_with_shipment.id,
+        shipment: [{ id: shipment.id }],
       )
-      expect(batch.num_shipments).to eq(0)
+      expect(batch_without_shipment.num_shipments).to eq(0)
     end
   end
 
   describe '.label' do
     it 'generates a label for a batch', :vcr do
-      batch = described_class.create(
+      batch = client.batch.create(
         shipments: [Fixture.one_call_buy_shipment],
       )
 
-      batch.buy
+      bought_batch = client.batch.buy(batch.id)
 
       unless File.file?(VCR.current_cassette.file)
         sleep(5) # Wait enough time for the batch to process buying the shipment
       end
 
-      batch.label(
-        file_format: 'ZPL',
-      )
+      batch_with_label = client.batch.label(bought_batch.id, { file_format: 'ZPL' })
 
       # We can't assert anything meaningful here because the label gets queued for generation and may not be immediately available
-      expect(batch).to be_an_instance_of(described_class)
+      expect(batch_with_label).to be_an_instance_of(EasyPost::Models::Batch)
     end
   end
 end