+++ spec/carrier_account_spec.rb
@@ -2,84 +2,79 @@
 
 require 'spec_helper'
 
-describe EasyPost::CarrierAccount, :authenticate_prod do
+describe EasyPost::Services::CarrierAccount do
+  let(:client) { EasyPost::Client.new(api_key: ENV['EASYPOST_PROD_API_KEY']) }
+
   describe '.create' do
     it 'creates a carrier account' do
-      carrier_account = described_class.create(Fixture.basic_carrier_account)
+      carrier_account = client.carrier_account.create(Fixture.basic_carrier_account)
 
-      expect(carrier_account).to be_an_instance_of(described_class)
+      expect(carrier_account).to be_an_instance_of(EasyPost::Models::CarrierAccount)
       expect(carrier_account.id).to match('ca_')
       expect(carrier_account.type).to eq('DhlEcsAccount')
 
       # Remove the carrier account once we have tested it so we don't pollute the account with test accounts
-      carrier_account.delete
+      client.carrier_account.delete(carrier_account.id)
     end
 
     it 'sends FedexAccount to the correct endpoint' do
-      allow(EasyPost).to receive(:make_request).with(
-        :post, '/v2/carrier_accounts/register', nil, { carrier_account: { type: 'FedexAccount' } },
+      allow(client).to receive(:make_request).with(
+        :post, 'carrier_accounts/register',
+        EasyPost::Models::CarrierAccount,
+        { carrier_account: { type: 'FedexAccount' } },
       ).and_return({ 'id' => 'ca_123' })
 
-      response = described_class.create(type: 'FedexAccount')
+      response = client.carrier_account.create(type: 'FedexAccount')
 
-      expect(response).to be_an_instance_of(described_class)
+      expect(response['id']).to eq('ca_123')
     end
   end
 
   describe '.retrieve' do
     it 'retrieves a carrier account' do
-      carrier_account = described_class.create(Fixture.basic_carrier_account)
-      retrieved_carrier_account = described_class.retrieve(carrier_account.id)
+      carrier_account = client.carrier_account.create(Fixture.basic_carrier_account)
+      retrieved_carrier_account = client.carrier_account.retrieve(carrier_account.id)
 
-      expect(retrieved_carrier_account).to be_an_instance_of(described_class)
+      expect(retrieved_carrier_account).to be_an_instance_of(EasyPost::Models::CarrierAccount)
       expect(retrieved_carrier_account.to_s).to eq(carrier_account.to_s)
 
       # Remove the carrier account once we have tested it so we don't pollute the account with test accounts
-      carrier_account.delete
+      client.carrier_account.delete(carrier_account.id)
     end
   end
 
   describe '.all' do
     it 'retrieves all carrier accounts' do
-      carrier_accounts = described_class.all
+      carrier_accounts = client.carrier_account.all
 
-      expect(carrier_accounts).to all(be_an_instance_of(described_class))
+      expect(carrier_accounts).to all(be_an_instance_of(EasyPost::Models::CarrierAccount))
     end
   end
 
-  describe '.save' do
+  describe '.update' do
     it 'updates a carrier account' do
       test_description = 'My custom description'
 
-      carrier_account = described_class.create(Fixture.basic_carrier_account)
+      carrier_account = client.carrier_account.create(Fixture.basic_carrier_account)
 
-      carrier_account.description = test_description
-      carrier_account.save
+      updated_carrier_account = client.carrier_account.update(carrier_account.id, { description: test_description })
 
-      expect(carrier_account).to be_an_instance_of(described_class)
-      expect(carrier_account.id).to match('ca_')
-      expect(carrier_account.description).to eq(test_description)
+      expect(updated_carrier_account).to be_an_instance_of(EasyPost::Models::CarrierAccount)
+      expect(updated_carrier_account.id).to match('ca_')
+      expect(updated_carrier_account.description).to eq(test_description)
 
       # Remove the carrier account once we have tested it so we don't pollute the account with test accounts
-      carrier_account.delete
+      client.carrier_account.delete(carrier_account.id)
     end
   end
 
   describe '.delete' do
     it 'deletes a carrier account' do
-      carrier_account = described_class.create(Fixture.basic_carrier_account)
+      carrier_account = client.carrier_account.create(Fixture.basic_carrier_account)
 
-      response = carrier_account.delete
-
-      expect(response).to be_an_instance_of(described_class)
-    end
-  end
-
-  describe '.types' do
-    it 'retrieves the carrier account types available' do
-      types = described_class.types
-
-      expect(types).to be_an_instance_of(Array)
+      expect {
+        client.carrier_account.delete(carrier_account.id)
+      }.not_to raise_error
     end
   end
 end