+++ spec/event_spec.rb
@@ -2,23 +2,25 @@
 
 require 'spec_helper'
 
-describe EasyPost::Event do
+describe EasyPost::Services::Event do
+  let(:client) { EasyPost::Client.new(api_key: ENV['EASYPOST_TEST_API_KEY']) }
+
   describe '.retrieve' do
     it 'retrieves an event' do
-      events = described_class.all(
+      events = client.event.all(
         page_size: Fixture.page_size,
       )
 
-      retrieved_event = described_class.retrieve(events.events[0].id)
+      retrieved_event = client.event.retrieve(events.events[0].id)
 
-      expect(retrieved_event).to be_an_instance_of(described_class)
+      expect(retrieved_event).to be_an_instance_of(EasyPost::Models::Event)
       expect(retrieved_event.id).to match('evt_')
     end
   end
 
   describe '.all' do
     it 'retrieves all events' do
-      events = described_class.all(
+      events = client.event.all(
         page_size: Fixture.page_size,
       )
 
@@ -26,23 +28,22 @@ describe EasyPost::Event do
 
       expect(events_array.count).to be <= Fixture.page_size
       expect(events.has_more).not_to be_nil
-      expect(events_array).to all(be_an_instance_of(described_class))
+      expect(events_array).to all(be_an_instance_of(EasyPost::Models::Event))
     end
   end
 
   describe '.get_next_page' do
     it 'retrieves the next page of a collection' do
-      first_page = described_class.all(
+      first_page = client.event.all(
         page_size: Fixture.page_size,
       )
 
       begin
-        next_page = described_class.get_next_page(first_page)
+        next_page = client.event.get_next_page(first_page)
 
         first_page_first_id = first_page.events.first.id
         next_page_first_id = next_page.events.first.id
 
-        # Did we actually get a new page?
         expect(first_page_first_id).not_to eq(next_page_first_id)
       rescue EasyPost::Error => e
         # If we get an error, make sure it's because there are no more pages.
@@ -53,9 +54,9 @@ describe EasyPost::Event do
 
   describe '.receive' do
     it 'receives (converts) an event' do
-      event = described_class.receive(Fixture.event_json)
+      event = EasyPost::Util.receive_event(Fixture.event_json)
 
-      expect(event).to be_an_instance_of(described_class)
+      expect(event).to be_an_instance_of(EasyPost::Models::EasyPostObject)
       expect(event.id).to match('evt_')
     end
   end
@@ -63,12 +64,12 @@ describe EasyPost::Event do
   describe '.retrieve_all_payloads' do
     it 'retrieve all payloads' do
       # Create a webhook to receive the event
-      webhook = EasyPost::Webhook.create(
+      webhook = client.webhook.create(
         url: Fixture.webhook_url,
       )
 
       # Create a batch to trigger an event
-      _ = EasyPost::Batch.create(
+      _ = client.batch.create(
         shipments: [Fixture.one_call_buy_shipment],
       )
 
@@ -76,31 +77,31 @@ describe EasyPost::Event do
         sleep(5) # Wait for the event to be created
       end
 
-      events = described_class.all(
+      events = client.event.all(
         page_size: Fixture.page_size,
       )
 
       event = events.events[0]
 
-      payloads = event.retrieve_all_payloads
+      payloads = client.event.retrieve_all_payloads(event.id)
 
       payload_array = payloads.payloads
 
-      expect(payload_array).to all(be_an_instance_of(EasyPost::Payload))
+      expect(payload_array).to all(be_an_instance_of(EasyPost::Models::Payload))
 
-      webhook.delete
+      client.webhook.delete(webhook.id)
     end
   end
 
   describe '.retrieve_a_payload' do
     it 'retrieve a payload' do
       # Create a webhook to receive the event
-      webhook = EasyPost::Webhook.create(
+      webhook = client.webhook.create(
         url: Fixture.webhook_url,
       )
 
       # Create a batch to trigger an event
-      _ = EasyPost::Batch.create(
+      _ = client.batch.create(
         shipments: [Fixture.one_call_buy_shipment],
       )
 
@@ -108,7 +109,7 @@ describe EasyPost::Event do
         sleep(5) # Wait for the event to be created
       end
 
-      events = described_class.all(
+      events = client.event.all(
         page_size: Fixture.page_size,
       )
 
@@ -116,12 +117,12 @@ describe EasyPost::Event do
 
       begin
         # Payload does not exist due to queueing, so this will throw an exception
-        event.retrieve_payload('payload_11111111111111111111111111111111')
+        client.event.retrieve_payload(event.id, 'payload_11111111111111111111111111111111')
       rescue EasyPost::Error => e
         expect(e.status).to eq(404)
       end
 
-      webhook.delete
+      client.webhook.delete(webhook.id)
     end
   end
 end