+++ spec/parcel_spec.rb
@@ -2,12 +2,14 @@
 
 require 'spec_helper'
 
-describe EasyPost::Parcel do
+describe EasyPost::Services::Parcel do
+  let(:client) { EasyPost::Client.new(api_key: ENV['EASYPOST_TEST_API_KEY']) }
+
   describe '.create' do
     it 'creates a parcel' do
-      parcel = described_class.create(Fixture.basic_parcel)
+      parcel = client.parcel.create(Fixture.basic_parcel)
 
-      expect(parcel).to be_an_instance_of(described_class)
+      expect(parcel).to be_an_instance_of(EasyPost::Models::Parcel)
       expect(parcel.id).to match('prcl_')
       expect(parcel.weight).to eq(15.4)
     end
@@ -15,17 +17,11 @@ describe EasyPost::Parcel do
 
   describe '.retrieve' do
     it 'retrieves a parcel' do
-      parcel = described_class.create(Fixture.basic_parcel)
-      retrieved_parcel = described_class.retrieve(parcel.id)
+      parcel = client.parcel.create(Fixture.basic_parcel)
+      retrieved_parcel = client.parcel.retrieve(parcel.id)
 
-      expect(retrieved_parcel).to be_an_instance_of(described_class)
+      expect(retrieved_parcel).to be_an_instance_of(EasyPost::Models::Parcel)
       expect(retrieved_parcel.to_s).to eq(parcel.to_s)
-    end
-  end
-
-  describe '.all' do
-    it 'raises not implemented error' do
-      expect { described_class.all }.to raise_error(NotImplementedError)
     end
   end
 end